fixed spelling of server.js
